pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "waam-research-rag"
version = "0.1.0"
description = "Production-style RAG service for WAAM and welding defect literature retrieval."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "OpenAI Codex" }
]
dependencies = [
  "fastapi>=0.115.0",
  "httpx>=0.28.0",
  "numpy>=2.0.0",
  "pydantic>=2.10.0",
  "PyYAML>=6.0.0",
  "python-json-logger>=3.2.0",
  "python-multipart>=0.0.9",
  "pypdf>=5.0.0",
  "scikit-learn>=1.5.0",
  "uvicorn[standard]>=0.34.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-cov>=5.0.0",
  "ruff>=0.8.0",
]
ml = [
  "sentence-transformers>=3.0.0",
  "torch>=2.4.0",
]
pdf-fallback = [
  "PyMuPDF>=1.24.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/waam_rag"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "-p no:cacheprovider"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "C4"]
